+++ CHANGELOG.md
@@ -44,3 +44,5 @@ All notable changes to this project will be documented in this file.
   - Small primary
   - Small secondary
 - Select widget.
+- Refactor User Component.
+- Reset select components in user component after saving a warning.